[WIP] Watch view implementation to track running 'odo watch' commands
Watch commands can be executed for different components at the same time
and that is done in terminal veiw. Over time terminal is getting
polluted with different commands and it is getting not easy to nawigate
and search for watch sessions to see the log or stop them.

This fix shows all components under watch in a view so devs can:
1. See the logs
2. Stop active watch commands

Related to #1425.

Signed-off-by: Denis Golovin <dgolovin@redhat.com>