Watch view implementation to track running 'odo watch' commands (#1612)
* [WIP] Watch view implementation to track running 'odo watch' commands

Watch commands can be executed for different components at the same time
and that is done in terminal veiw. Over time terminal is getting
polluted with different commands and it is getting not easy to navigate
and search for watch sessions to see the log, stop or check if watch is running already.

This fix shows all components under watch in a view so devs can:
1. See the logs
2. Stop active watch commands
3. Navigate to log if component watch is running already

Related to #1294.