Get rid of save workspace request in the end of integration tests
When integration tests are finished and vscode is about to exit
it shows 'Save Workspace Changes' request. That holds vscode form
exit and requires manually pushing skip button to finish tests.
To avoid this `.code-workspace` file should be used instead.
When tests end workspace should have the same projects it started
with and that lets vscode to close without any requests.

Signed-off-by: Denis Golovin dgolovin@redhat.com