Fix double refresh when creating the project
This PR fixes the issue with double refresh when create new project.
Creating new project triggers update of application view because of
changes in kubeconfig. At the same time create new project command
itself triggers tree update. That ends with tree updated, but error
appears in debug output about three item requested to be revealed
does not exists in the tree, because it was refreshed as result of
kubeconfig changes.

Signed-off-by: Denis Golovin dgolovin@redhat.com