POC for `odo push -o json` and icon animation in App Explorer
This PR uses `odo push -o json` to push component and showes in
terminal executed devfile commands and their output. Stdout output
is colored black and stderror is red.

This PR uses pseudo terminal to print output to terminal view, that
allows catch begining and end of command for specific node and
update icon and text to different ones while command is in progress.

Signed-off-by: Denis Golovin dgolovin@redhat.com