Use original devfile.yaml if present in context folder
This PR fixes #1815.

Wnen devfile is present in the root of the project, there
is no reason to ask for component type, because devfile is
component type. All is required is to call odo command
without type name and pass `--devfile devfile.yaml` option.
That slashes new component workflow - when ran from
Application View -  to just two steps. Selecting workspace
folder as context and typing in the component's name.
Default value for component name input is set as a name from
devfile.yaml.

Signed-off-by: Denis Golovin dgolovin@redhat.com