Fix folder selection for component
This fix removes the context folder selection step in case of
folder opened in workspace and use it as context.

If workspace has only one none component folder added to it
it is used as context.

If workspace has more than one folder added and only one is not
a compmponent the context folder selection step is still
showed to the user.

Signed-off-by: Denis Golovin dgolovin@redhat.com