Hide commands that require selecting a component
The commands are still accessible in the relevant right-click menus,
but are no longer available in the command palette,
since they don't work when invoked from the command palette.

This is a work around until we can get the commands to work from the
command palette when working with OpenShift Dev Sandbox.

Fixes #2792

Signed-off-by: David Thompson <davthomp@redhat.com>