If cluster is not up during login, pick new URL
When the user is attempting to login to a cluster,
check if the cluster is accessible before prompting for credentials.
To check if the cluster is accessible,
send a GET request to the cluster API URL;
if the API endpoint cannot be found:
- if the URL appears to be a CRC api URL, offer to start CRC
- if the URL appears to be a Dev Sandbox URL, link to the information page on Dev Sandbox
- otherwise, display a warning message and return to the step where the user inputs the cluster URL.

This means that if the user tries to login to a cluster that is not running,
they don't have to click through the rest of the inputs
before receiving this message.
Also, if the cluster is not available since the user mistyped the URL,
they have the chance to type it in again.

Fixes #709

Signed-off-by: David Thompson <davthomp@redhat.com>