Replace "delete component" command
The command "delete component" currently doesn't work
when run from the command palette and doesn't appear anywhere else.

In this PR, I have replaced it with two commands that you can run though
right clicking on a component in the component view.

The first command is "Delete Configuration Files".
This command runs `odo delete component --files -f`,
which removes the `devfile.yaml` and the `.odo` folder from the user's
computer,
so that OpenShift Toolkit no longer recognizes the project as an
OpenShift component.
It doesn't delete any of the source code from the user's computer.

The second command is "Delete Source Folder".
This command deletes the folder that the source code for a component is
in.

Some of the other features that used to be handled by "delete" can be
accomplished using other existing features:
- Stop Dev: if you are running the component in dev mode, `odo` will
  delete all the cluster resources associated with the component
  when you stop dev mode.
- Undeploy: if you have deployed the component to the cluster using `odo
  deploy` or the "Deploy" context menu, then you can use the "Undeploy"
  context menu to delete the resources associated with the component

Closes #2825

Signed-off-by: David Thompson <davthomp@redhat.com>