Add a new panel for all OpenShift commands
It's a webview, since we need to use either a treeview or webview for all
panels.
It wraps xtermjs, the same library VS Code uses for its terminal.
It exposes an API for interacting with the terminal multiplexer, as well
as with the individual running processes.
For example, we can run `odo dev` and update the UI
(like we were doing before this PR).

What's missing:
- [ ] copy/paste
- [ ] Tests (pending on https://github.com/redhat-developer/vscode-extension-tester/pull/855)
- [ ] reordering the tabs (I think we should gauge interest and save this for a future PR)
- [ ] tab bar scrolling (supposedly the Material UI component comes with
      this built it, but it wasn't working when I tested it)

Signed-off-by: David Thompson <davthomp@redhat.com>