Switch from webpack to esbuild for bundling webviews
This PR uses esbuild instead of webpack to bundle the webviews.
This dramatically reduces the time it takes to bundle the webviews,
from >50 seconds to <1 second.

This PR also modifies the npm scripts and VS Code launch configurations
to always recompile the webviews before launching.

I also set up eslint to be run before launching the extension.
If the extension seems slower to launch in debug mode, this is why.

Closes #2875

Signed-off-by: David Thompson <davthomp@redhat.com>