If a command is rerun after exiting, close the old tab
When you start a new command in the terminal,
check if there is a tab open where that command was run and has exited.
If there is, close that tab before opening an new tab to run the
rerun of the command.
Not quite the same as recycling the tab, but it has a much simpler
implementation.

Closes #3138

Signed-off-by: David Thompson <davthomp@redhat.com>