Reduce memory usage of build process
This takes the esbuild memory usage from 5 Gb to 1.8 Gb,
for both the command line run and the background build process for
VS Code.
It accomplishes this by doing everything in one esbuild call instead of
many parallel esbuild calls.
The tradeoff is that the directory structure of the `out` folder has
changed,
so I had to change some webview code to accomidate this change.

Signed-off-by: David Thompson <davthomp@redhat.com>