Make esbuild to produce bundled and minified VSIX
This fix makes 'esbuild' to produce the bundled and minified VSIX extension archive,
free of unneeded dependency modules, when the packaging is started with `vsce package`.

On other hand, when building with `npm install && npm run build` and testing with `npm test`
the extension file structure is kept unchange and the trunspiled scripts aren't minified,
so the unit testing and coverage tests work as usual.

Note: `npm install` is needed to be executed after executing `vsce package` as the last one
clears the `node_modules/` of the depencencies not needed in production.

Fixes #4226

Signed-off-by: Victor Rubezhny <vrubezhny@redhat.com>