Cannot login to existing cluster #3935
The `openshiftToolkit.failOnBrokenKubeConfigEntry` (boolean) option is added with the
default set to `false` in order to make Kube config loading more robust.
If it's required to fully check the correctness of the Kube config on startup, users can
set its value to `true`, so Kubernetes Client will throw errors if any.

Fixes: #3935

Signed-off-by: Victor Rubezhny <vrubezhny@redhat.com>